error handling in mutations, removed unnecessary traits
